requirements.txt
aiohttp==3.9.3
pyyaml==6.0.1
numpy==1.24.4
requests==2.31.0
tqdm==4.66.2
scipy==1.10.1
jsonschema==4.21.1
